Creating modules for the various parts
- Created a module for the Book class
- Created a module for the Methods class
- Created a module for the DOM elements
- Installed the luxon library
