+++ modules/BookClass.js
@@ -1,0 +1,8 @@
+// Define Book Class
+export default class Book {
+  constructor(title, author, id) {
+    this.title = title;
+    this.author = author;
+    this.id = id;
+  }
+}
