+++ modules/DOMElements.js
@@ -1,0 +1,16 @@
+const form = document.querySelector('#form');
+const ul = document.querySelector('#book-list');
+const bookSection = document.querySelector('#bookSection');
+const heading = document.querySelector('.page-title');
+const dateTime = document.querySelector('#time');
+const msg = document.querySelector('.no-books');
+
+// Getting sections of The HTML
+const contactSection = document.querySelector('#contact');
+const contactLink = document.querySelector('#contact-link');
+const list = document.querySelector('#list');
+const addNew = document.querySelector('#add');
+
+export {
+  form, ul, bookSection, heading, dateTime, msg, contactSection, contactLink, list, addNew,
+};
